+++ My Showcase/src/store/modules/clickerGame/settings.js
@@ -1,0 +1,4 @@
+export const settings = {
+    saveInterval: 60000,
+    debug: false,
+}